requirements.txt
sanic~=22.6.2
psycopg2-binary~=2.9.3